requirements/common.txt
# Core Stuff
# -------------------------------------
Django==2.1

# Configuration
# -------------------------------------
argon2-cffi==18.1.0
django-environ==0.4.5
django-sites==0.10
python-dotenv==0.9.1
django-cors-headers==2.2.0



# Extensions
# -------------------------------------
pytz==2018.3

# Models
# -------------------------------------
psycopg2-binary==2.7.4

Pillow==5.0.0
django-extensions==2.0.7
django-uuid-upload-path==1.0.0
django-versatileimagefield==1.9

# REST APIs
# -------------------------------------
djangorestframework==3.8.2
django-rest-swagger==2.2.0

# LOGGING
# -------------------------------------
django-log-request-id==1.3.2

# Documentation
# -------------------------------------
mkdocs==1.0



# Auth Stuff
# -------------------------------------
PyJWT==1.6.4
django-mail-templated==2.6.5
